use higher range for postgre

+++ students/models.py
@@ -3,7 +3,7 @@ from teachers.models import Subjects
 # Create your models here.
 
 class Student(models.Model):
-    student_id = models.IntegerField(primary_key=True)
+    student_id = models.BigIntegerField(primary_key=True)
     first_name = models.CharField(max_length=100, default='')
     last_name = models.CharField(max_length=100, default='')
     email = models.EmailField(null=True)
